Rename submitter folder to orchestrator
